requirements.txt
numpy==2.4.0
paddleocr==3.3.2
pandas==2.3.3
pdf2image==1.17.0
Pillow==12.0.0
tqdm==4.67.1
